gang: Simple initial implementation
This has all the factors pointing in the correct direction, but
there are plenty of tunables which could still be optimized if
wanted.

Much of the delay will still come from getting your karma low enough
to make a gang in the first place, though. This is enough to grow and
dominate the field within about a day of the gang being founded, with
minimal external revenue.