omni: fix softlock
If all reserved normalization threads were in use, a replan triggered,
and the new normalization process saw that all threads were in use,
it would not reschedule itself on the new queue, incorrectly determining
that because threads were in use, a schedule_normalize was already in
(or would later be added to) the queue.

To fix this, also keep track of the normalization threads used by this
planner in particular. It's a bit ugly, but would need a rethink to
replace it with something cleaner.