aug: Implement priorities
Still does not yet account for prereqs, but the new reinsertion and
recalculation logic will get most of the way.

Also does not account for faction reputation. This is more awkwards to
do optimally, as you can save money based on which factions you buy
from, but there are a few obvious heuristics from which we can probably
put together something which is good enough.